+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 - 1.0.8
   - Added support for functions as value source
+  - Improved Mongo format: now can compare with field & use funcs with help of `$expr`
+  - (breaking) `mongoFormatOp` - inserted arg `useExpr` at 4th position
 - 1.0.7
   - Fixed Babel
 - 1.0.6